+++ src/Exceptions/RavenContextNotFoundException.php
@@ -4,7 +4,7 @@ namespace ChijiokeIbekwe\Raven\Exceptions;
 
 use Exception;
 
-class RavenEntityNotFoundException extends Exception
+class RavenContextNotFoundException extends Exception
 {
     protected $code = 404;
 }